# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/api/v1/events/{id}/notification-preference" : {
      "get" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Preference"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "set_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Preference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Preference"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/telegram/webhook" : {
      "post" : {
        "tags" : [ "telegram-webhook-controller" ],
        "operationId" : "receive",
        "parameters" : [ {
          "name" : "X-Telegram-Bot-Api-Secret-Token",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : ""
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "format" : "byte"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{id}/comments" : {
      "get" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "section",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "GENERAL"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "create",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CommentInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string",
                    "format" : "uuid"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comments/{id}" : {
      "delete" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      },
      "patch" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "edit_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Edit"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/me/notifications" : {
      "get" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Preference" : {
        "type" : "object",
        "properties" : {
          "level" : {
            "type" : "string"
          }
        }
      },
...
      "CommentInput" : {
        "type" : "object",
        "properties" : {
          "section" : {
            "type" : "string"
          },
          "parentCommentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "body" : {
            "type" : "string"
          }
        }
      },
...
      "Edit" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string"
          }
        }
      },
...
      "Comment" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "authorId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "authorName" : {
            "type" : "string"
          },
          "section" : {
            "type" : "string"
          },
          "parentCommentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "editedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deletedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}